Move Token type to auth module
